# specs/001-todo-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Backend API
  version: 1.0.0
  description: Multi-user Todo API with JWT authentication via Better Auth JWKS

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /api/health:
    get:
      operationId: healthCheck
      summary: Health check
      tags: [Health]
      security: []
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                required: [status]
        "503":
          description: Service unhealthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [unhealthy]
                  detail:
                    type: string
                required: [status, detail]

  /api/todos:
    post:
      operationId: createTodo
      summary: Create a new todo
      tags: [Todos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoCreate"
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoPublic"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

    get:
      operationId: listTodos
      summary: List authenticated user's todos
      tags: [Todos]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoPublic"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      operationId: updateTodo
      summary: Update a todo
      tags: [Todos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoUpdate"
      responses:
        "200":
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoPublic"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      operationId: deleteTodo
      summary: Delete a todo
      tags: [Todos]
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Todo deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by Better Auth, verified via JWKS (RS256)

  schemas:
    TodoCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 2000
          nullable: true
        is_completed:
          type: boolean
          default: false

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 2000
          nullable: true
        is_completed:
          type: boolean

    TodoPublic:
      type: object
      required: [id, title, is_completed, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        oneOf:
                          - type: string
                          - type: integer
                    msg:
                      type: string
                    type:
                      type: string